Registering package dependencies within the package
This removes the need for the user to manually register these themselves, ensuring they only need to register one service provider (i.e. for this package).

In this case the package dependency is a more generic kind, likely to be used by one than one package and maybe the app itself. This is a test to see if the same package can be registered twice, or if Laravel won't like that (in which such packages should be defined as they are now in app/config/app.php).